fix: remove old batch.go to fix redeclaration errors
The new batch_channel.go (now batch.go) replaces the old implementation with:
- No savepointMu lock needed
- SAVEPOINT at Commit time
- Simplified design
